refactor: 주소 검색 컴포넌트 수정
